+++ delphi-frontend/src/test/resources/au/com/integradev/delphi/grammar/NamedArguments.pas
@@ -1,0 +1,13 @@
+unit NamedArguments;
+
+interface
+
+implementation
+
+procedure Test(Foo: OleVariant);
+begin
+  Foo.Bar(Baz := 'Flarp');
+  Foo.Bar(string := 123);
+end;
+
+end.